changes to hardware branch
